Moving scan playback logic out of EDUS2View.
